[Refactor]: Improved testing and middleware in 1st exercise part4

+++ part4/bloglist/controllers/blogs.js
@@ -1,6 +1,7 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/Blog')
 const User = require('../models/User')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -8,11 +9,11 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.post('/', async (request, response, next) => {
-  const { title, author, url, likes } = request.body
+  let { title, author, url, likes } = request.body
 
-  const { userId } = request
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-  const user = await User.findById(userId)
+  const user = await User.findById(decodedToken.id)
 
   if (!likes) {
     likes = 0
@@ -38,9 +39,21 @@ blogsRouter.post('/', async (request, response, next) => {
   }
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', async (request, response, next) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  try {
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if (blogToDelete.user.toString() === decodedToken.id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(401).end()
+    }
+  } catch (error) {
+    next(error)
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
